+++ src/_igraph/convert.h
@@ -58,6 +58,7 @@ int igraphmodule_PyObject_to_attribute_combination_type_t(PyObject* o,
 int igraphmodule_PyObject_to_barabasi_algorithm_t(PyObject *o,
   igraph_barabasi_algorithm_t *result);
 int igraphmodule_PyObject_to_bliss_sh_t(PyObject *o, igraph_bliss_sh_t *result);
+int igraphmodule_PyObject_to_coloring_greedy_t(PyObject *o, igraph_coloring_greedy_t *result);
 int igraphmodule_PyObject_to_community_comparison_t(PyObject *obj,
                   igraph_community_comparison_t *result);
 int igraphmodule_PyObject_to_connectedness_t(PyObject *o, igraph_connectedness_t *result);